+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.13.1] - 2026-01-26
+
+### Added
+
+- Added custom archive format
+
+### Fixed
+
+- Fixed recover command
+
 ## [0.13.0] - 2025-11-04
 
 ### Changed
@@ -161,6 +171,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release
 
+[0.13.1]: https://github.com/ShrBox/BackupHelper/compare/v0.13.0...v0.13.1
 [0.13.0]: https://github.com/ShrBox/BackupHelper/compare/v0.12.1...v0.13.0
 [0.12.1]: https://github.com/ShrBox/BackupHelper/compare/v0.12.0...v0.12.1
 [0.12.0]: https://github.com/ShrBox/BackupHelper/compare/v0.11.0...v0.12.0